# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Employee time tracking SaaS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type EmployeeStatus =
  (typeof EmployeeStatus)[keyof typeof EmployeeStatus];

export const EmployeeStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface Employee {
  id: number;
  name: string;
  email: string;
  department: string;
  position: string;
  /** Expected check-in time HH:MM */
  expectedCheckin: string;
  /** Expected check-out time HH:MM */
  expectedCheckout: string;
  status: EmployeeStatus;
  createdAt: string;
}

export type CreateEmployeeBodyStatus =
  (typeof CreateEmployeeBodyStatus)[keyof typeof CreateEmployeeBodyStatus];

export const CreateEmployeeBodyStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface CreateEmployeeBody {
  name: string;
  email: string;
  department: string;
  position: string;
  expectedCheckin: string;
  expectedCheckout: string;
  status?: CreateEmployeeBodyStatus;
}

export type UpdateEmployeeBodyStatus =
  (typeof UpdateEmployeeBodyStatus)[keyof typeof UpdateEmployeeBodyStatus];

export const UpdateEmployeeBodyStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface UpdateEmployeeBody {
  name?: string;
  email?: string;
  department?: string;
  position?: string;
  expectedCheckin?: string;
  expectedCheckout?: string;
  status?: UpdateEmployeeBodyStatus;
}

export type AttendanceRecordStatus =
  (typeof AttendanceRecordStatus)[keyof typeof AttendanceRecordStatus];

export const AttendanceRecordStatus = {
  open: "open",
  closed: "closed",
} as const;

export interface AttendanceRecord {
  id: number;
  employeeId: number;
  employeeName: string;
  employeeDepartment: string;
  date: string;
  clockIn: string;
  clockOut?: string | null;
  totalMinutes?: number | null;
  status: AttendanceRecordStatus;
  notes?: string | null;
  /** Minutes late compared to expected checkin */
  lateMinutes?: number | null;
  overtimeMinutes?: number | null;
}

export interface CreateAttendanceBody {
  employeeId: number;
  clockIn: string;
  clockOut?: string;
  notes?: string;
}

export interface UpdateAttendanceBody {
  clockIn?: string;
  clockOut?: string;
  notes?: string;
}

export interface ClockInBody {
  employeeId: number;
}

export interface ClockOutBody {
  employeeId: number;
}

export interface DashboardSummary {
  totalEmployees: number;
  presentToday: number;
  absentToday: number;
  lateToday: number;
  clockedOutToday: number;
  stillWorkingNow: number;
  averageHoursThisWeek: number;
  attendanceRateThisMonth: number;
}

export type DailyHoursEntryStatus =
  (typeof DailyHoursEntryStatus)[keyof typeof DailyHoursEntryStatus];

export const DailyHoursEntryStatus = {
  present: "present",
  absent: "absent",
  holiday: "holiday",
  weekend: "weekend",
} as const;

export interface DailyHoursEntry {
  date: string;
  clockIn?: string | null;
  clockOut?: string | null;
  hoursWorked?: number | null;
  status: DailyHoursEntryStatus;
  lateMinutes?: number | null;
}

export interface EmployeeHoursReport {
  employeeId: number;
  employeeName: string;
  month: number;
  year: number;
  totalDays: number;
  presentDays: number;
  absentDays: number;
  lateDays: number;
  totalHours: number;
  expectedHours: number;
  overtimeHours: number;
  dailyBreakdown: DailyHoursEntry[];
}

export interface DepartmentSummary {
  department: string;
  totalEmployees: number;
  presentToday: number;
  absentToday: number;
  lateToday: number;
  attendanceRate: number;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  clock_in: "clock_in",
  clock_out: "clock_out",
} as const;

export interface ActivityItem {
  id: number;
  employeeId: number;
  employeeName: string;
  department: string;
  type: ActivityItemType;
  timestamp: string;
  lateMinutes?: number | null;
}

export interface LateArrivalItem {
  employeeId: number;
  employeeName: string;
  department: string;
  expectedCheckin: string;
  actualCheckin: string;
  lateMinutes: number;
}

export type ListEmployeesParams = {
  search?: string;
  department?: string;
  status?: ListEmployeesStatus;
};

export type ListEmployeesStatus =
  (typeof ListEmployeesStatus)[keyof typeof ListEmployeesStatus];

export const ListEmployeesStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export type ListAttendanceParams = {
  employeeId?: number;
  dateFrom?: string;
  dateTo?: string;
  status?: ListAttendanceStatus;
};

export type ListAttendanceStatus =
  (typeof ListAttendanceStatus)[keyof typeof ListAttendanceStatus];

export const ListAttendanceStatus = {
  open: "open",
  closed: "closed",
} as const;

export type GetEmployeeHoursReportParams = {
  month?: number;
  year?: number;
};

export type GetRecentActivityParams = {
  limit?: number;
};