# lib/api-zod/src/generated/types/employee.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Employee time tracking SaaS API
 * OpenAPI spec version: 0.1.0
 */
import type { EmployeeStatus } from "./employeeStatus";

export interface Employee {
  id: number;
  name: string;
  email: string;
  department: string;
  position: string;
  /** Expected check-in time HH:MM */
  expectedCheckin: string;
  /** Expected check-out time HH:MM */
  expectedCheckout: string;
  status: EmployeeStatus;
  createdAt: Date;
}