# lib/api-zod/src/generated/types/updateEmployeeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Employee time tracking SaaS API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateEmployeeBodyStatus } from "./updateEmployeeBodyStatus";

export interface UpdateEmployeeBody {
  name?: string;
  email?: string;
  department?: string;
  position?: string;
  expectedCheckin?: string;
  expectedCheckout?: string;
  status?: UpdateEmployeeBodyStatus;
}